Added imperfect interface to multimaterial contact, but not for 3D yet (though will run) and not all tested yet. Also added volume-weighted gradient method (and made it the default) for normals in multimaterial contact.

+++ Common/System/DataTypes.hpp
@@ -65,7 +65,7 @@ enum { XX=0,YY,ZZ,YZ,XZ,XY};
 	
 	// contact law properties, nextFriction and matID only used for material contact
 	typedef struct {
-		short law;
+		int law;
 		double friction;
 		double Dn;
 		double Dnc;
